Accurate and fast cell marker gene identification with COSG

Marker gene identification for cell groups in a given dataset.

cosg(
  object,
  groups = "all",
  assay = "RNA",
  slot = "data",
  mu = 1,
  n_genes_user = 100
)

Arguments

assay
Assay to use in marker gene identification

slot
Slot to pull data from

mu
The penalty factor to penalize gene expression in cells not belonging to the cluster of interest

n_genes_user
Number of top ranked genes returned in the result

Value

A list containing two dataframes for ranked marker genes' names and scores, respectively

Examples

suppressMessages(library(Seurat))
data('pbmc_small',package='Seurat')
# Check cell groups:
table(Idents(pbmc_small))

#> 
#>  0  1  2 
#> 36 25 19 

#######
# Run COSG:
marker_cosg <- cosg(
 pbmc_small,
 groups='all',
 assay='RNA',
 slot='data',
 mu=1,
 n_genes_user=100)
#######
# Check the marker genes:
 head(marker_cosg$names)

#>       0      1     2
#> 1   CD7 S100A8 MS4A1
#> 2  CCL5   TYMP CD79A
#> 3  GNLY S100A9 TCL1A
#> 4 LAMP1  FCGRT  NT5C
#> 5  GZMA IFITM3 CD79B
#> 6   LCK   LST1 FCER2

 head(marker_cosg$scores)

#>           0         1         2
#> 1 0.6391917 0.8954042 0.6922908
#> 2 0.6391267 0.8312083 0.5832425
#> 3 0.6328148 0.8120045 0.5757478
#> 4 0.6164937 0.7755955 0.5533107
#> 5 0.5846589 0.7413060 0.5163446
#> 6 0.5795238 0.7380483 0.5115180